rename "service" stuff to "provider"
- to keep to the standard
