+++ src/pipes/account-label/account-label.ts
@@ -1,18 +1,18 @@
 import { Pipe, PipeTransform } from '@angular/core';
 import { UserDataProvider } from '@providers/user-data/user-data';
-import { ContactsService } from '@providers/contacts/contacts.service';
+import { ContactsProvider } from '@providers/contacts/contacts';
 
 @Pipe({
   name: 'accountLabel',
 })
 export class AccountLabelPipe implements PipeTransform {
 
-  constructor(public userDataProvider: UserDataProvider, private contactsService: ContactsService) {
+  constructor(public userDataProvider: UserDataProvider, private contactsProvider: ContactsProvider) {
 
   }
 
   transform(value: string, defaultText: string, ...args) {
-    const contact = this.contactsService.getContactByAddress(value);
+    const contact = this.contactsProvider.getContactByAddress(value);
     if (contact) { return contact.name; }
 
     const wallet = this.userDataProvider.getWalletByAddress(value);
